# openapi.yaml
openapi: 3.1.0
info:
  title: 27001certified API
  version: 1.0.0
  description: |
    Backend API for 27001certified.app — ISO 27001:2022 AI-Certification Portal.
    All protected endpoints require a JWT Bearer token unless marked as public.
  contact:
    name: 27001certified Platform
    url: https://27001certified.app
servers:
  - url: https://api.27001certified.app
    description: Production
  - url: https://api-staging.27001certified.app
    description: Staging
  - url: http://localhost:5000
    description: Local development

# ────────────────────────────────────────────────────────────
#  SECURITY SCHEMES
# ────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SHARED SCHEMAS ─────────────────────────────────────────
  schemas:
...
    User:
      type: object
      properties:
        id:            { type: string, format: uuid }
        full_name:     { type: string }
        email:         { type: string, format: email }
        company:       { type: string, nullable: true }
        job_role:      { type: string, nullable: true }
        country:       { type: string, nullable: true }
        email_verified:{ type: boolean }
        role:          { type: string, enum: [user, admin] }
        created_at:    { type: string, format: date-time }
...
    ExamSession:
      type: object
      properties:
        id:            { type: string, format: uuid }
        user_id:       { type: string, format: uuid }
        track:         { type: string, enum: [associate, professional] }
        status:
          type: string
          enum: [generating, ready, active, completed, expired, aborted]
        score:         { type: integer, nullable: true }
        passed:        { type: boolean, nullable: true }
        tab_violations:{ type: integer }
        started_at:    { type: string, format: date-time, nullable: true }
        completed_at:  { type: string, format: date-time, nullable: true }
        expires_at:    { type: string, format: date-time, nullable: true }
        created_at:    { type: string, format: date-time }
    Certificate:
      type: object
      properties:
        id:               { type: string, format: uuid }
        certificate_id:   { type: string, example: "ISO27-2026-A1B2C3D4" }
        user_id:          { type: string, format: uuid }
        exam_session_id:  { type: string, format: uuid }
        track:            { type: string, enum: [associate, professional] }
        full_name:        { type: string }
        score:            { type: integer }
        awarded_on:       { type: string, format: date-time }
        r2_url_landscape: { type: string, format: uri, nullable: true }
        r2_url_square:    { type: string, format: uri, nullable: true }
        revoked:          { type: boolean }
        revoked_reason:   { type: string, nullable: true }
...
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: string }

    # ── Pagination ─────────────────────────────────────────
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page:  { type: integer }
        pages: { type: integer }
        limit: { type: integer }

  # ── RESPONSES ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - BearerAuth: []

# ────────────────────────────────────────────────────────────
#  PATHS
# ────────────────────────────────────────────────────────────
paths:
...
  /api/user/me:
    get:
      summary: Get current user profile + attempt summary
      tags: [User]
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/User' }
                  - type: object
                    properties:
                      attempts_associate:   { type: integer }
                      attempts_professional:{ type: integer }
                      last_attempt_at:      { type: string, format: date-time, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      summary: Update profile fields
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                company:   { type: string }
                job_role:  { type: string }
                country:   { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      summary: Soft-delete account (GDPR)
      tags: [User]
      responses:
        '200':
          description: Account scheduled for deletion
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/user/attempts:
    get:
      summary: Get paginated list of past exam sessions
      tags: [User]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10, maximum: 50 }
        - in: query
          name: track
          schema: { type: string, enum: [associate, professional] }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Pagination' }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ExamSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ══════════════════════════════════════════
  #  SESSION (Exam Generation + Exam Taking)
  # ══════════════════════════════════════════
...
  /api/certificates/{certId}:
    get:
      summary: Get certificate (owner only)
      tags: [Certificates]
      parameters:
        - in: path
          name: certId
          required: true
          schema: { type: string }
          description: Certificate ID (ISO27-2026-XXXXXXXX) or UUID
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Certificate' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/certificates/verify/{certId}:
    get:
      summary: Public certificate verification (no auth required)
      security: []
      tags: [Certificates]
      parameters:
        - in: path
          name: certId
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:        { type: boolean }
                  certificate:
                    type: object
                    description: Public fields only — full_name and score are omitted to protect candidate privacy
                    properties:
                      certificate_id: { type: string }
                      track:          { type: string, enum: [associate, professional] }
                      issued_at:      { type: string, format: date-time }
                      valid_until:    { type: string, format: date-time, nullable: true }
                      status:         { type: string, enum: [active, revoked] }
        '404': { $ref: '#/components/responses/NotFound' }

  # ══════════════════════════════════════════
  #  ADMIN — all require role=admin
  # ══════════════════════════════════════════
...